Attempt at pagination editing index.html
The {{ content }} line in default.html in fact refers to the content in
index.html. What generates the list of posts in the News tab is
index.html, therefore the paginator mechanism should be placed there.
I still don't know if the paths created will be correct as I'm not
editing the paginate_path variable.